# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Custom Analytic - WW2 API",
        "description": "API Documentation for Custom Analytic - WW2",
        "contact": {
            "email": "flazen.edu@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/facilities/summary": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "summary": "Show facilities summary",
                "operationId": "bbab88a2e5f981c36b7559c70fe398d1",
                "responses": {
                    "200": {
                        "description": "facilities found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/facilities/total/bytype/{limit}": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "summary": "Total facilities by type",
                "operationId": "bc7cd01fe97871b63ea2a54cb1796572",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of facility per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "facilities found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/facilities/total/bycountry/{limit}": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "summary": "Total facilities by country",
                "operationId": "a0e2ed5390e24afad15f90c038cbf623",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "path",
                        "description": "Number of facility per page",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "facilities found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/facilities/total/bylocation/{type}": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "summary": "Total facilities by location and type",
                "operationId": "a17158ce2d2ff0f3bec482858397015d",
                "parameters": [
                    {
                        "name": "type",
                        "in": "path",
                        "description": "Filter by field type. Can be facility type or 'NULL'",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Airfield"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "facilities found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/facilities/total/bysides": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "summary": "Total facilities by side",
                "operationId": "12373d25ef9f3acc2e4b96c445565da7",
                "responses": {
                    "200": {
                        "description": "facilities found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/facilities/type": {
            "get": {
                "tags": [
                    "Facility"
                ],
                "summary": "Show all facility type",
                "operationId": "d5e419ea3ce6f1cfd15c60f76698fc07",
                "responses": {
                    "200": {
                        "description": "facilities found"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "something wrong. please contact admin"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}